+++ metview/bindings.py
@@ -21,7 +21,7 @@ import cffi
 import numpy as np
 
 
-__version__ = "1.5.2"
+__version__ = "1.6.0"
 
 
 def string_from_ffi(s):